# internal/order/ports/openapi_types.gen.go
// Package ports provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.4.1 DO NOT EDIT.
package ports

// CreatOrderRequest defines model for CreatOrderRequest.
type CreatOrderRequest struct {
	CustomerID string             `json:"customerID"`
	Items      []ItemWithQuantity `json:"items"`
}

// Error defines model for Error.
type Error struct {
	Message *string `json:"message,omitempty"`
}

// Item defines model for Item.
type Item struct {
	Id          *string `json:"id,omitempty"`
	Name        *string `json:"name,omitempty"`
	PaymentLink *string `json:"paymentLink,omitempty"`
	PriceID     *string `json:"priceID,omitempty"`
	Quantity    *int32  `json:"quantity,omitempty"`
}

// ItemWithQuantity defines model for ItemWithQuantity.
type ItemWithQuantity struct {
	Id       *string `json:"id,omitempty"`
	Quantity *int32  `json:"quantity,omitempty"`
}

// Order defines model for Order.
type Order struct {
	CustomerID  *string `json:"customerID,omitempty"`
	Id          *string `json:"id,omitempty"`
	Items       *[]Item `json:"items,omitempty"`
	PaymentLink *string `json:"paymentLink,omitempty"`
	Status      *string `json:"status,omitempty"`
}

// GetCustomerCustomerIDOrdersJSONRequestBody defines body for GetCustomerCustomerIDOrders for application/json ContentType.
type GetCustomerCustomerIDOrdersJSONRequestBody = CreatOrderRequest